Extract issue meta surface projection helper
